initial support for article searching (thanks bernimoses)
due to the fact that Redmine requires a project association for this to work,
we've had to add the project model as a dependency of article. We also need to
have the article's project_id value match a VALID project, which has status of 1.

TODO: hack support for ensuring that the kb_articles.project_id points to a project
with a projects.status = 1